third_party/WebKit/LayoutTests/http/tests/websocket/handshake-fail-by-null-subprotocol-selection.html

<!DOCTYPE HTML>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description("Test for verifying that the WebSocket handshake fails when the client requests a subprotocol but the server doesn't include a Sec-WebSocket-Protocol header in the response.");

window.jsTestIsAsync = true;

var url = "ws://localhost:8880/handshake-fail-by-null-subprotocol-selection";
var ws = new WebSocket(url, 'test');
var closeEvent;

ws.onopen = function()
{
    testFailed("Connection established.");
    ws.close();
};

ws.onclose = function(event)
{
    closeEvent = event;
    shouldBeFalse("closeEvent.wasClean");
    finishJSTest();
};

ws.onerror = function(errorEvent)
{
    testPassed("onerror() was called");
};

</script>
</body>
</html>
