third_party/blink/web_tests/http/tests/websocket/handshake-fail-by-null-subprotocol-selection-expected.txt

Test for verifying that the WebSocket handshake fails when the client requests a subprotocol but the server doesn't include a Sec-WebSocket-Protocol header in the response.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS onerror() was called
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
